Made test installation and split boot test into long and short
