+++ bin/test
@@ -1,3 +1,4 @@
+#!/bin/bash
 #==============================================================================
 # Copyright (C) 2017 Stephen F. Norledge and Alces Software Ltd.
 #
@@ -19,14 +20,5 @@
 # For more information on the Alces Metalware, please visit:
 # https://github.com/alces-software/metalware
 #==============================================================================
-require_relative "#{ENV['alces_BASE']}/test/helper/base-test-require.rb" 
-$: << "#{ENV['alces_BASE']}/test"
-
-require 'capture-test'
-require 'iterator-test'
-require 'kickstart-test'
-require 'hosts-test'
-require 'boot-test-quick'
-require 'boot-test-long'
-require 'templater-test/combiner-test'
-require 'templater-test/finder-test'
+$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )/exportCmd
+alias rby=$alces_RUBY
